requirements.txt
absl-py==2.2.2
asgiref==3.8.1
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Django==5.2.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.0
fsspec==2025.3.2
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.31.4
idna==3.10
Jinja2==3.1.6
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.9
networkx==3.4.2
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
timm==1.0.15
torch==2.7.0+cpu
torchvision==0.22.0+cpu
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.140
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.17.2
